requirements.txt
attrs==21.4.0
build==0.8.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
hypothesis==6.47.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pathlib==1.0.1
pep517==0.12.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sortedcontainers==2.4.0
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
XlsxWriter==3.0.3
zipp==3.8.0
